+++ src/components/Filter.tsx
@@ -1,16 +1,18 @@
 import React from 'react'
-import { IAction, IPlayReply, TFilterEvents } from '../reducers/initialState'
-import { ActionDescriptor } from 'photoshop/dist/types/photoshop'
-import { action } from '../imports'
 import './Filter.css'
+import { TFilterEvents } from '../reducers/initialState'
 
 export interface IFilterProps{
-
+	filterType: TFilterEvents
+	include: string
+	exclude: string
 }
 
 export interface IFilterDispatch{
 	setSearchTerm(str: string): void
-	setFilterEventsType(type:TFilterEvents):void
+	setFilterEventsType(type: TFilterEvents): void
+	setInclude(arr:string[]):void
+	setExclude(arr:string[]):void
 }
 
 type TFilter = IFilterProps & IFilterDispatch
@@ -24,9 +26,38 @@ export default class Filter extends React.Component<TFilter> {
 		this.props.setSearchTerm(e.currentTarget.value);
 	}
 
-	private onSetFilterEventsType = (e:any,type:TFilterEvents) => {
-		debugger;
-		this.props.setFilterEventsType(e)
+	private onSetFilterEventsType = (e: any) => {
+		this.props.setFilterEventsType(e.target.value)
+	}
+
+	private setExclude = (e: React.ChangeEvent<HTMLInputElement>) => {
+		this.props.setExclude(e.currentTarget.value.split(";"));
+	}
+
+	private setInclude = (e: React.ChangeEvent<HTMLInputElement>) => {
+		this.props.setInclude(e.currentTarget.value.split(";"));
+	}
+
+	private renderFilterFields = () => {
+		switch (this.props.filterType) {
+			case "exclude": {
+				return (
+					<React.Fragment>
+						<span>Exclude: </span><input onChange={this.setExclude} value={this.props.exclude} type="text" />
+					</React.Fragment>
+				)
+			}
+			case "include": {
+				return (
+					<React.Fragment>
+						<span>Include: </span><input onChange={this.setInclude} value={this.props.include} type="text" />
+					</React.Fragment>
+				);
+			}
+			case "none": {
+				return null
+			}
+		}
 	}
 
 	public render() {
@@ -34,11 +65,19 @@ export default class Filter extends React.Component<TFilter> {
 			<React.Fragment>
 				<div className="FilterComponent">
 					<span>Search: </span><input onChange={this.onSearch} type="text" />
-
-
-					{/*<input onChange={this.onSearch} type="text" />*/}
+					<span>Filter:</span>
+					<sp-dropdown>
+						<sp-menu slot="options" onClick={this.onSetFilterEventsType}>
+							<sp-menu-item value="none">None</sp-menu-item>
+							<sp-menu-item value="include">Include</sp-menu-item>
+							<sp-menu-item value="exclude">Exclude</sp-menu-item>
+						</sp-menu>
+					</sp-dropdown>
+				</div>
+				<div className="ExcludeInclude">
+					{this.renderFilterFields()}
 				</div>
 			</React.Fragment>
-		)
+		);
 	}
 }