+++ src/components/FilterContainer.tsx
@@ -1,7 +1,7 @@
 import { connect } from 'react-redux'
 import { IAppState, IAction } from '../reducers/initialState'
-import { getCollapsedDefault, getBatchPlayDecorator } from '../selectors'
-import { toggleCollapseOptionAction, addReplyAction, toggleExpandAction, setSearchTermAction } from '../actions/actions'
+import { getCollapsedDefault, getBatchPlayDecorator, getFilterType, getInclude, getExclude } from '../selectors'
+import { toggleCollapseOptionAction, addReplyAction, toggleExpandAction, setSearchTermAction, setFilterTypeAction, setIncludeAction, setExcludeAction } from '../actions/actions'
 import { IFilterProps, IFilterDispatch } from './Filter'
 import cloneDeep from "lodash/cloneDeep"
 import Filter from './Filter'
@@ -9,14 +9,18 @@ import Filter from './Filter'
 
 const mapStateToProps = (state: IAppState): IFilterProps => {
 	return {
-
+		filterType: getFilterType(state),
+		include: getInclude(state),
+		exclude: getExclude(state)
 	}
 }
 
 const mapDispatchToProps = (dispatch: any):IFilterDispatch => {
 	return {
 		setSearchTerm: (str) => dispatch(setSearchTermAction(str)),
-		setFilterEventsType:(type)=>dispatch()
+		setFilterEventsType: (type) => dispatch(setFilterTypeAction(type)),
+		setInclude:(arr)=>dispatch(setIncludeAction(arr)),
+		setExclude:(arr)=>dispatch(setExcludeAction(arr)),
 	}
 }